+++ RestaurantApp.Client/src/app/components/product/product.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder, Validators, NgForm } from '@angular/forms';
 import { Product } from 'src/app/_models/product';
 import { ProductService } from 'src/app/_services/product.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-product',
@@ -12,8 +13,9 @@ export class ProductComponent implements OnInit {
   products: Product[] = [];
   form: FormGroup;
   submitted: boolean = false;
+  usage: boolean = false;
   
-  constructor(private service: ProductService, private fb: FormBuilder) { }
+  constructor(private toastr: ToastrService, private service: ProductService, private fb: FormBuilder) { }
 
   ngOnInit() {
     this.service.getProducts().subscribe(res => this.products = res);
@@ -42,6 +44,7 @@ export class ProductComponent implements OnInit {
         this.products.push(res);
         form.reset();
         this.submitted = false;
+        this.toastr.success('You have been inserted the product successfully.', 'Successfully');
       },
       err => {
         console.log(err);
@@ -51,18 +54,28 @@ export class ProductComponent implements OnInit {
   }
 
   onDelete(product: Product): void {
-    if (confirm('Are you sure to delete ' + product.productName + '?')) {
-      this.service.deleteProduct(product).subscribe(
-        res => {
-          const index = this.products.indexOf(product);
-          this.products.splice(index, 1);
-        },
-        err => {
-          console.log(err);
-          alert(err);
+    this.service.getProductUsage(product.id).subscribe(
+      res => {
+        this.usage = res;
+        if (this.usage === true) {
+          this.toastr.error('You cannot delete ' +  product.productName + ' since it is using on an order.', 'Error');
+        } else {
+          if (confirm('Are you sure to delete ' + product.productName + '?')) {
+            this.service.deleteProduct(product).subscribe(
+              res => {
+                const index = this.products.indexOf(product);
+                this.products.splice(index, 1);
+                this.toastr.warning('You have been deleted ' + product.productName + ' successfully.', 'Successfully');
+              },
+              err => {
+                console.log(err);
+                alert(err);
+              }
+            );
+          }
         }
-      );
-    }
+      }
+    );
   }
 
 }
